added new app/view/db & template
