+++ post/templates/post/posts.html
@@ -1,0 +1,30 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block content %}
+<div class="container-fluid">
+    <h1 class="my-4">Blog Posts</h1>
+    <div class="row">
+        {% for post in posts %}
+            <div class="col-md-4">
+                <div class="card mb-4 p-md-3 rounded">
+                    <div class="card-body">
+                        <div class="image-container">
+                            {% if post.image %}
+                                <img class="card-img-top" src="{{ post.image.url }}" alt="{{ post.author_name }}">
+                            {% else %}
+                                <img class="card-img-top" src="/static/images/placeholder.jpg" alt="placeholder image">
+                            {% endif %}
+                        </div>  
+                        <p class="card-text">{{ post.body }}</p>
+                        <p class="author">By {{ post.author_name }}</p>
+                        <p class="date">{{ post.posted_date }}</p>
+                    </div>
+                </div>
+            </div>
+        {% empty %}
+            <p>No posts available.</p>
+        {% endfor %}
+    </div>
+</div>
+{% endblock %}
