+++ post/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views 
+
+urlpatterns = [
+    path('', views.PostListView.as_view(), name='posts'),
+]
