--- templates/post.html
@@ -1,62 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<div class="container-fluid">
-    <div class="row">
-
-        <!-- Blog Entries Column -->
-        <div class="col-12 mt-3 contain">
-            <div class="post-title">
-                <h1 class="lead-text my-5">My hikes</h1>
-                <form method="GET" action="{% url 'search_results' %}">
-                    {% csrf_token %}
-                    <div class="form-parent my-4 mx-5">
-                        <input type="search" name="query" placeholder="Search post..." class="form-box ws" required>
-                        <button type="submit" class="search-btn"><i class="fas fa-search"></i></button>
-                    </div>
-                </form>
-            </div>
-
-            <div class="row height m-md-2">
-                <div class="col-md-4">
-                    <div class="card mb-4 p-md-3 rounded">
-                        <div class="card-body">
-                            <!-- If we are using image -->
-                            <div class="image-container">
-                                {% if "placeholder" in post.featured_image.url %}
-                                <img class="card-img-top" src="/static/images/placeholder.jpg" alt="placeholder image">
-                                {% else %}
-                                <img class="card-img-top" src="" alt="post-image">
-                                {% endif %}
-                            </div>
-                             <!--Image container -->
-                            <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-                            <p class="author"> Signed name</p>
-                            
-                        </div>
-                    </div>
-                </div>
-                {% if forloop.counter|divisibleby:3 %}
-            </div>
-            <div class="row height m-md-2">
-                {% endif %}
-                {% endfor %}
-
-            </div>
-            {% if is_paginated %}
-            <nav aria-label="Page navigation">
-                <ul class="pagination justify-content-center">
-                    {% if page_obj.has_previous %}
-                    <li><a href="?page={{ page_obj.previous_page_number }}" class="page-link btn-header">&laquo; PREV </a></li>
-                    {% endif %}
-                    {% if page_obj.has_next %}
-                    <li ><a href="?page={{ page_obj.next_page_number }}" class="page-link btn-header"> NEXT &raquo;</a></li>
-
-                    {% endif %}
-                </ul>
-            </nav>
-            {% endif %}
-        </div>
-    </div>
-</div>
-{% endblock content %}
